SwiftCompilerSources: split the Optimizer/Context.swift file
