+++ SwiftCompilerSources/Sources/Optimizer/PassManager/CMakeLists.txt
@@ -7,9 +7,10 @@
 # See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 
 swift_compiler_sources(Optimizer
-  Context.swift
+  ContextCommon.swift
+  FunctionPassContext.swift
   ModulePassContext.swift
   Options.swift
   Passes.swift
-  PassRegistration.swift)
-
+  PassRegistration.swift
+  SimplifyContext.swift)
